+++ pairlist.cpp
@@ -2,7 +2,7 @@
 
 // ============================================
 
-void set_pair(Pair &pair, int i, int j, int idi, int idj) {
+void set_pair(Pair &pair, unsigned long i, unsigned long j, unsigned long idi, unsigned long idj) {
     pair.i = i;
     pair.j = j;
     pair.idi = idi;
@@ -17,11 +17,11 @@ void PairList::make_pair(Variables* &vars, Systemparam* sysp) {
     // 自領域内ペア
     list.clear();
     Atom *atoms = vars->atoms.data();
-    const int pn = vars->number_of_atoms();
-    for (int i=0; i<pn; i++) {
+    const unsigned long pn = vars->number_of_atoms();
+    for (unsigned long i=0; i<pn; i++) {
         double ix = atoms[i].x;
         double iy = atoms[i].y;
-        for (int j=i+1; j<pn; j++) {
+        for (unsigned long j=i+1; j<pn; j++) {
             double dx = atoms[j].x - ix;
             double dy = atoms[j].y - iy;
             periodic_distance(dx, dy, sysp);
@@ -44,16 +44,16 @@ void PairList::make_pair(Variables* &vars, Systemparam* sysp) {
     std::vector<std::vector<Atom>> new_other_atom;
     for (auto &one_other_atoms : vars->other_atoms) {
         Atom *other_atoms = one_other_atoms.data();
-        const int other_pn = one_other_atoms.size();
+        const unsigned long other_pn = one_other_atoms.size();
         std::vector<Pair> one_other_list;
         std::vector<Atom> one_new_other_atom;
-        std::vector<int> one_recv_list;
-        int pair_j_index = 0;   // other_atomを間引いてしまうので、ペアのインデックスは新しく用意
-        for (int i=0; i<other_pn; i++) {
+        std::vector<unsigned long> one_recv_list;
+        unsigned long pair_j_index = 0;   // other_atomを間引いてしまうので、ペアのインデックスは新しく用意
+        for (unsigned long i=0; i<other_pn; i++) {
             double ix = other_atoms[i].x;
             double iy = other_atoms[i].y;
             bool survive = false;
-            for (int j=0; j<pn; j++) {
+            for (unsigned long j=0; j<pn; j++) {
                 double dx = atoms[j].x - ix;
                 double dy = atoms[j].y - iy;
                 periodic_distance(dx, dy, sysp);
